# contracts/recomendacao-escolas.openapi.yaml
openapi: 3.0.3
info:
  title: Recomendação de Escola
  version: "2.0.0"
  description: >
    Recomenda escolas para uma família, combinando proximidade (moradia/trabalho),
    adequação entre o score estimado da família e o quão concorrida cada escola é
    historicamente, e o histórico de não-comparecimento do responsável. Consumido
    pela Tela 1 (Inscrição). Não depende de nenhum outro módulo em runtime — score e
    histórico chegam prontos no corpo do pedido (ver contracts/README.md).
paths:
  /health:
    get:
      summary: Liveness check
      responses:
        "200":
          description: OK
  /cep/{cep}:
    get:
      summary: "Resolve um CEP para bairro e coordenada"
      description: >
        Resolve pela tabela local dados/ceps.csv (sem rede). Só cai no ViaCEP
        (viacep.com.br) para CEP fora dessa tabela — e nesse caso latitude/longitude
        voltam nulas, porque o ViaCEP não devolve coordenada.
      parameters:
        - name: cep
          in: path
          required: true
          schema:
            type: string
          description: "CEP com ou sem hífen (ex.: 23050-300 ou 23050300)."
      responses:
        "200":
          description: Bairro e, quando conhecida, a coordenada do CEP
          content:
            application/json:
              schema:
                type: object
                required: ["bairro"]
                properties:
                  bairro: { type: string }
                  latitude:
                    type: number
                    nullable: true
                    description: "Nula quando o CEP só foi resolvido pelo ViaCEP."
                  longitude:
                    type: number
                    nullable: true
                    description: "Nula quando o CEP só foi resolvido pelo ViaCEP."
        "404":
          description: "CEP inválido, inexistente, ou ViaCEP indisponível"
  /escolas:
    get:
      summary: "Lista simples de escolas de um bairro, sem ranking (ex.: para um mapa)"
      parameters:
        - name: bairro
          in: query
          required: true
          schema:
            type: string
          description: Nome do bairro (com ou sem acento, qualquer capitalização).
      responses:
        "200":
          description: Lista de escolas encontradas (pode ser vazia)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./schemas/escola.schema.json"
        "422":
          description: Parâmetro bairro ausente ou vazio
  /recomendacoes:
    post:
      summary: Recomenda e explica as melhores escolas para um pedido
      parameters:
        - name: limite
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
          description: Quantidade máxima de escolas retornadas (a escola_preferida, se informada, sempre é incluída além do limite).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/pedido_recomendacao.schema.json"
      responses:
        "200":
          description: Escolas recomendadas, ordenadas por pontuacao_final (desc)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./schemas/recomendacao_escola.schema.json"
        "422":
          description: "Pedido inválido (ex.: nenhum endereço informado)"
  /admin/recomendacoes:
    get:
      summary: Formulário HTML para testar/auditar o recomendador manualmente
      responses:
        "200":
          description: Página HTML com o formulário
          content:
            text/html:
              schema:
                type: string
    post:
      summary: Processa o formulário e renderiza a tabela de recomendações com o racional
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                bairro_moradia: { type: string }
                bairro_trabalho: { type: string }
                percentil_score: { type: string, description: "0-100, vazio = sem score" }
                vezes_convocado: { type: string }
                vezes_nao_compareceu: { type: string }
                escola_preferida_esc_codigo: { type: string }
      responses:
        "200":
          description: Página HTML com a tabela de recomendações e o racional de cada uma
          content:
            text/html:
              schema:
                type: string